requirements.txt
aniso8601==9.0.1
attrs==21.2.0
click==8.0.1
ffmpeg-python==0.2.0
Flask==2.0.1
Flask-Caching==1.10.1
flask-restplus==0.13.0
flask-swagger-ui==3.36.0
Flask-Uploads==0.2.1
future==0.18.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonschema==3.2.0
MarkupSafe==2.0.1
numpy==1.21.1
opencv-contrib-python==4.5.3.56
opencv-python==4.5.3.56
Pillow==8.3.1
pyrsistent==0.18.0
pytz==2021.1
scikit-video==1.1.11
scipy==1.7.0
six==1.16.0
Werkzeug==2.0.1
